Introduce IterableOnceOps as common abstraction for Iterable & Iterator
- Unlike the other Ops types, IterableOnceOps is *not* implemented by
  IterableOnce but only by Iterator and Iterable.

- Fix the hierarchy of MultiSets and SortedMultiSets: The refactoring
  caused soundness problems to show up that were not reported before.
  These types now follow the usual diamond pattern where
  mutable/immutable extends generic and sorted extends unsorted.
