+++ collections-contrib/src/main/scala/strawman/collection/immutable/MultiSet.scala
@@ -10,10 +10,38 @@ import strawman.collection.decorators.ImmutableMapDecorator
   * An immutable multiset
   * @tparam A the element type of the collection
   */
-class MultiSet[A] private (elems: Map[A, Int])
+trait MultiSet[A]
   extends collection.MultiSet[A]
     with Iterable[A]
-    with collection.MultiSetOps[A, MultiSet, MultiSet[A]] {
+    with MultiSetOps[A, MultiSet, MultiSet[A]] {
+
+  override def iterableFactory: IterableFactory[MultiSet] = MultiSet
+}
+
+trait MultiSetOps[A, +CC[X] <: MultiSet[X], +C <: MultiSet[A]] extends collection.MultiSetOps[A, CC, C] {
+  /**
+    * @return an immutable multiset containing all the elements of this multiset
+    *         and one more occurrence of `elem`
+    * @param elem the element to add
+    */
+  def incl(elem: A): C
+
+  /** Alias for `incl` */
+  @`inline` final def + (elem: A): C = incl(elem)
+
+  /**
+    * @return an immutable multiset containing all the elements of this multiset
+    *         and one occurrence less of `elem`
+    *
+    * @param elem the element to remove
+    */
+  def excl(elem: A): C
+
+  /** Alias for `excl` */
+  @`inline` final def - (elem: A): C = excl(elem)
+}
+
+class MultiSetImpl[A] private[immutable] (elems: Map[A, Int]) extends MultiSet[A] {
 
   def occurrences: Map[A, Int] = elems
 
@@ -25,13 +53,10 @@ class MultiSet[A] private (elems: Map[A, Int])
     * @param elem the element to add
     */
   def incl(elem: A): MultiSet[A] =
-    new MultiSet(elems.updatedWith(elem) {
+    new MultiSetImpl(elems.updatedWith(elem) {
       case None    => Some(1)
       case Some(n) => Some(n + 1)
     })
-
-  /** Alias for `incl` */
-  @`inline` final def + (elem: A): MultiSet[A] = incl(elem)
 
   /**
     * @return an immutable multiset containing all the elements of this multiset
@@ -40,12 +65,9 @@ class MultiSet[A] private (elems: Map[A, Int])
     * @param elem the element to remove
     */
   def excl(elem: A): MultiSet[A] =
-    new MultiSet(elems.updatedWith(elem) {
+    new MultiSetImpl(elems.updatedWith(elem) {
       case Some(n) => if (n > 1) Some(n - 1) else None
     })
-
-  /** Alias for `excl` */
-  @`inline` final def - (elem: A): MultiSet[A] = excl(elem)
 
 }
 
@@ -57,7 +79,7 @@ object MultiSet extends IterableFactory[MultiSet] {
       case _ => (newBuilder[A]() ++= source).result()
     }
 
-  def empty[A] = new MultiSet[A](Map.empty)
+  def empty[A] = new MultiSetImpl[A](Map.empty)
 
   def newBuilder[A](): Builder[A, MultiSet[A]] =
     new ImmutableBuilder[A, MultiSet[A]](empty[A]) {
