+++ collections-contrib/src/main/scala/strawman/collection/mutable/MultiSet.scala
@@ -7,13 +7,16 @@ import strawman.collection.decorators.MutableMapDecorator
 /**
   * A mutable multiset.
   */
-class MultiSet[A] private (val elems: Map[A, Int])
+trait MultiSet[A]
   extends collection.MultiSet[A]
     with collection.MultiSetOps[A, MultiSet, MultiSet[A]]
     with Growable[A]
     with Shrinkable [A] {
 
   override def iterableFactory: IterableFactory[MultiSet] = MultiSet
+}
+
+class MultiSetImpl[A] private[mutable] (val elems: Map[A, Int]) extends MultiSet[A] {
 
   def occurrences: collection.Map[A, Int] = elems
 
@@ -39,7 +42,7 @@ object MultiSet extends IterableFactory[MultiSet] {
 
   def from[A](source: IterableOnce[A]): MultiSet[A] = (newBuilder[A]() ++= source).result()
 
-  def empty[A]: MultiSet[A] = new MultiSet[A](Map.empty)
+  def empty[A]: MultiSet[A] = new MultiSetImpl[A](Map.empty)
 
   def newBuilder[A](): Builder[A, MultiSet[A]] = new GrowableBuilder[A, MultiSet[A]](empty)
